Make Sass Spec npm installable
Node Sass to have Sass Spec specs as dependency for inclusion in
the node-citgm. The requirement for being the citgm is that a
packages tests should be executable via `npm install`.

This means Node Sass needs to publish it's tests to npm. That
includes Sass Spec. At the moment Sass Spec is a git submodule so
we'd need to publish the entire repo. This is a [problem][1] as
Sass Spec weighs in at 69M.

Size aside some users have experienced issues with some unicode
[file names][2] and [long paths][3]. To get around these issues
we want remove the git submodule and instead `npm install` the
Sass Spec spec files. We'd do this a `devDependency` so regular
users won't install them which avoids the previously mentioned issues.
Having only the specs files in npm package will also hugely reduce
the package size.

[1]: https://twitter.com/nodkz/status/773103622678798336
[2]: https://github.com/sass/node-sass/issues/1699
[3]: https://github.com/sass/node-sass/issues/1701